refactor spectrogram FFT
